+++ IELTSGlanceWidget/IELTSGlanceWidgetBundle.swift
@@ -1,0 +1,16 @@
+//
+//  IELTSGlanceWidgetBundle.swift
+//  IELTSGlanceWidget
+//
+//  Created by severushou on 2026/7/16.
+//
+
+import WidgetKit
+import SwiftUI
+
+@main
+struct IELTSGlanceWidgetBundle: WidgetBundle {
+    var body: some Widget {
+        IELTSGlanceWidget()
+    }
+}